refactor(domain): 자동차 움직이는 기능 di를 활용해서 리팩토링
