+++ src/main/java/racingcar/domain/Car.java
@@ -1,14 +1,15 @@
-package racingcar.model;
+package racingcar.domain;
 
 public class Car {
     private CarName carName;
     private int currentPosition;
+
     public Car(CarName carName) {
         this.carName = carName;
     }
 
-    public void move(RandomMoveNumber randomMoveNumber) {
-        if(randomMoveNumber.okToGo()) {
+    public void move(MovingStrategy movingStrategy) {
+        if (movingStrategy.movable()) {
             currentPosition++;
         }
     }
@@ -24,7 +25,7 @@ public class Car {
 
     private String getCurrentPositionString() {
         StringBuilder currentPositionString = new StringBuilder();
-        for(int i = 0; i < currentPosition; i++) {
+        for (int i = 0; i < currentPosition; i++) {
             currentPositionString.append("-");
         }
         return currentPositionString.toString();
