+++ src/main/java/racingcar/domain/MovingStrategy.java
@@ -1,0 +1,6 @@
+package racingcar.domain;
+
+@FunctionalInterface
+public interface MovingStrategy {
+    boolean movable();
+}
