+++ src/main/java/racingcar/domain/RandomMoveNumber.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 public class RandomMoveNumber {
     private int value;
