+++ src/test/java/racingcar/domain/RacingGameTest.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 import org.junit.jupiter.api.Test;
 import racingcar.util.CarUtil;
@@ -12,19 +12,14 @@ class RacingGameTest {
 
     @Test
     void 입력받은값으로_자동차객체들_생성() {
-        RacingGame racingGame = new RacingGame();
+        RacingGame racingGame = new RacingGame(() -> true);
         String carNameString = "asd,zxc,qwe";
         racingGame.createCars(CarUtil.createCarsOf(carNameString));
     }
 
     @Test
     void 사용자입력값에따라_자동차들이_움직인다() {
-        RacingGame racingGame = new RacingGame() {
-            @Override
-            protected RandomMoveNumber getRandomMoveNumber() {
-                return new RandomMoveNumber(4);
-            }
-        };
+        RacingGame racingGame = new RacingGame(() -> true);
 
         String carNamesString = "asd,zxc,qwe";
         List<Car> cars = CarUtil.createCarsOf(carNamesString);
